requirements.txt
langchain
cloudinary==1.33.0
firebase_admin==6.1.0
nltk==3.8.1
openai
Pillow
protobuf==4.23.3
python-slugify
Requests
#spacy
stability_sdk
streamlit
textblob
python-dotenv
grpcio-tools
google-cloud-translate
googletrans==4.0.0-rc1
#simpleaichat
